+++ src/classes/AuthStatusEnum.js
