+++ src/classes/CurrentUser.js
