+++ src/components/pages/admin/Admin_EditQuestionWeights.jsx
@@ -3,9 +3,9 @@ import "./css/Admin_EditQuestionWeights.css";
 import React, { useState, useEffect } from "react";
 import { useLoaderData } from "react-router-dom";
 import { Popup } from "reactjs-popup";
-import MenuBar from './MenuBar';
-import AWS_Authenticator from './AWS_Authenticator';
-import AuthStatusEnum from "./AuthStatusEnum";
+import MenuBar from '../../../MenuBar';
+import AWS_Authenticator from '../../../AWS_Authenticator';
+import AuthStatusEnum from "../../AuthStatusEnum";
 
 function Admin_EditQuestionWeights() {
   var departments = useLoaderData();
