+++ src/components/pages/auth/Login.jsx
@@ -7,7 +7,7 @@
 import minesbkgd from './images/mines-bkgd.jpg';
 import './css/Portal.css';
 import { Amplify } from 'aws-amplify';
-import awsconfig from './aws-exports';
+import awsconfig from '../../../aws-exports';
 import { useState, useEffect } from 'react';
 import { signIn } from '@aws-amplify/auth';
 import { FaEyeSlash, FaEye } from 'react-icons/fa';
