+++ src/components/pages/professor/Professor_Home.jsx
@@ -1,8 +1,8 @@
 import './css/Professor_Home.css';
-import MenuBar from './MenuBar';
-import SurveyCard from './SurveyCard';
+import MenuBar from '../../../MenuBar';
+import SurveyCard from '../../../SurveyCard';
 import { Amplify } from 'aws-amplify';
-import awsconfig from './aws-exports';
+import awsconfig from '../../../aws-exports';
 import AWS_Authenticator from './AWS_Authenticator';
 import AuthStatusEnum from './AuthStatusEnum';
