+++ src/components/pages/student/Student_TakeSurvey.jsx
@@ -1,7 +1,7 @@
 /* frontend display of the survey. This displays the acutal content of the survey. A student gets here by clicking on the survey */
 // By: Wesley Woo
 
-import MenuBar from './MenuBar';
+import MenuBar from '../../../MenuBar';
 import './css/Student_TakeSurvey.css';
 import React, { useState } from 'react';
 import AWS_Authenticator from './AWS_Authenticator';
